(#1091) Fix tests around zocalo triggering
